server/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
Pillow==10.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.13.2